+++ src/test/java/test/org/fugerit/java/script/helper/TestEvalScript.java
@@ -20,7 +20,7 @@ import java.util.function.Function;
 @Slf4j
 class TestEvalScript {
 
-    private EvalScript evalKts = new EvalScriptWithJsonDataModel( "kts" );
+    private EvalScript evalKts = EvalScriptWithJsonDataModel.newEvalScriptWithJsonDataModel( "kts" );
 
     private String testWorkerSimpleScript(Function<Reader, Object> evalFun) throws IOException {
         try (InputStreamReader reader = new InputStreamReader(ClassHelper.loadFromDefaultClassLoader( "kts/sample-1.kts" ))) {
@@ -33,8 +33,8 @@ class TestEvalScript {
 
     @Test
     void testEvalkts() throws IOException {
-        Assertions.assertNotNull( this.testWorkerSimpleScript( r -> this.evalKts.evalKts( r ) ) );
-        Assertions.assertNotNull( this.testWorkerSimpleScript( r -> SafeFunction.get( () -> this.evalKts.evalKtsEx( r ) ) ) );
+        Assertions.assertNotNull( this.testWorkerSimpleScript( r -> this.evalKts.handle( r ) ) );
+        Assertions.assertNotNull( this.testWorkerSimpleScript( r -> SafeFunction.get( () -> this.evalKts.handleEx( r ) ) ) );
     }
 
     @Test
@@ -42,7 +42,7 @@ class TestEvalScript {
         try (InputStreamReader reader = new InputStreamReader(ClassHelper.loadFromDefaultClassLoader( "kts/sample-1-with-data-model.kts" ))) {
             Map<String, Object> dataModel = new HashMap<>();
             dataModel.put( "docTitle", "My custom title" );
-            Object result = evalKts.evalKts( reader, dataModel );
+            Object result = evalKts.handle( reader, dataModel );
             String xml = result.toString();
             log.info( "xml from data model : \n{}", xml );
             Assertions.assertNotNull( xml );
@@ -52,7 +52,7 @@ class TestEvalScript {
     @Test
     void testEvalktsWithException() throws IOException {
         try (InputStreamReader reader = new InputStreamReader(ClassHelper.loadFromDefaultClassLoader( "kts/sample-1-with-data-model.kts" ))) {
-            Assertions.assertThrows( ScriptException.class, () -> evalKts.evalKtsEx( reader ) ) ;
+            Assertions.assertThrows( ScriptException.class, () -> evalKts.handleEx( reader ) ) ;
         }
     }
 
@@ -62,6 +62,11 @@ class TestEvalScript {
         Assertions.assertNotNull( new ScriptException( new IOException() ) );
         Assertions.assertNotNull( new ScriptException( "test 1" ) );
         Assertions.assertNotNull( new ScriptException( "test 2", new ConfigException()) );
+    }
+
+    @Test
+    void testConstructors() {
+        Assertions.assertNotNull( EvalScriptWithJsonDataModel.newEvalScriptWithJsonDataModel( "kts", "custom" ) );
     }
 
 }