+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- EvalScript method evalKts() changed in handle()
+- EvalScriptWithJsonDataModel divided in EvalScriptWithDataModel and EvalScriptWithJsonDataModel
+
 ## [1.0.0] - 2024-10-23
 
 ### Added